refactor: PalletCmd run with spec
