+++ crates/pop-cli/src/commands/mod.rs
@@ -5,8 +5,6 @@ use clap::Subcommand;
 use pop_common::templates::Template;
 use serde_json::{json, Value};
 
-use sp_runtime::traits::BlakeTwo256;
-
 pub(crate) mod bench;
 pub(crate) mod build;
 pub(crate) mod call;
@@ -15,11 +13,6 @@ pub(crate) mod install;
 pub(crate) mod new;
 pub(crate) mod test;
 pub(crate) mod up;
-
-type HostFunctions = (
-	sp_statement_store::runtime_api::HostFunctions,
-	cumulus_primitives_proof_size_hostfunction::storage_proof_size::HostFunctions,
-);
 
 #[derive(Subcommand)]
 #[command(subcommand_required = true)]
@@ -96,22 +89,7 @@ impl Command {
 				},
 			},
 			#[cfg(feature = "parachain")]
-			Self::Bench(args) => match args.command {
-				#[cfg(feature = "parachain")]
-				bench::Command::Pallet(cmd) => {
-					if let Some(spec) = cmd.shared_params.chain {
-						return Err(anyhow::anyhow!(format!(
-							"Chain specs are not supported. Please remove `--chain={spec}` and use \
-								`--runtime=<PATH>` instead"
-						)));
-					}
-					cmd.run_with_spec::<BlakeTwo256, HostFunctions>(None)
-						.map_err(|_| {
-							anyhow::anyhow!(format!("Failed to run benchmarking for the pallet"))
-						})
-						.map(|_| Value::Null)
-				},
-			},
+			Self::Bench(args) => bench::Command::execute(args).map(|_| Value::Null),
 			#[cfg(any(feature = "parachain", feature = "contract"))]
 			Self::Build(args) => match args.command {
 				None => build::Command::execute(args).map(|t| json!(t)),
